destructure builder
